requirements.txt
Flask==1.1.2
graphql-core>=3
aniso8601==8.1.1
